updated jupyter general image
The image is now based on , which has libraries for data analysis and manipulation. We were already installing most of the libraries ourselves, so it seems sensible to use this pre-baked image. I also removed all of the old dependencies from previous semesters, and just installed the ones that were requested on top of what is provided by the new base image. Last but not least, I added dependencies and configuration options to support printing to pdf via html